rockiot/requirements.txt
django
djangorestframework
djangorestframework-gis
djangorestframework-simplejwt
django-admin-interface
django-import-export
psycopg2-binary
configparser
python-decouple
pyrabbit2
pika
paho-mqtt
dataclasses-json
docker
requests
celery
kombu
django_celery_results
django_celery_beat
drf-yasg
sqlalchemy==1.3.23
django-simple-history
flower
gunicorn
whitenoise
django-leaflet
django-geojson[field]